Extended Browser Support and cleanup
