+++ pom.xml
@@ -4,7 +4,7 @@
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <modelVersion>4.0.0</modelVersion>
 
-    <groupId>org.sneha.shashank</groupId>
+    <groupId>org.bitnervee</groupId>
     <artifactId>Automation_Core-Framework</artifactId>
     <version>1.0-SNAPSHOT</version>
