+++ src/main/java/com/shashank/initialization/BrowserFactory.java
@@ -1,0 +1,47 @@
+package com.shashank.initialization;
+
+import io.github.bonigarcia.wdm.WebDriverManager;
+import org.openqa.selenium.WebDriver;
+import org.openqa.selenium.chrome.ChromeDriver;
+import org.openqa.selenium.edge.EdgeDriver;
+import org.openqa.selenium.firefox.FirefoxDriver;
+import org.openqa.selenium.safari.SafariDriver;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+import org.testng.annotations.AfterClass;
+import org.testng.annotations.BeforeClass;
+
+public class BrowserFactory{
+
+    public static final Logger logger = LoggerFactory.getLogger(BrowserFactory.class);
+
+    WebDriver driver = null;
+    @BeforeClass
+    public void setUp(){
+        String browser = System.getProperty("browser","safari");
+        logger.debug("Selecting selenium browser using" );
+        if (browser.contains("chrome")){
+            WebDriverManager.chromedriver().setup();
+            driver = new ChromeDriver();
+        }else if (browser.contains("safari")){
+            WebDriverManager.safaridriver().setup();
+            driver = new SafariDriver();
+        }else if (browser.contains("firefox")){
+            WebDriverManager.firefoxdriver().setup();
+            driver = new FirefoxDriver();
+        }else {
+            WebDriverManager.edgedriver().setup();
+            driver = new EdgeDriver();
+        }
+
+
+        PageDriver.setDriver(driver);
+    }
+
+
+
+    @AfterClass
+    public void tearDown(){
+        PageDriver.getDriver().quit();
+    }
+}
